Win32 Parallel port code upgraded to work better on machines without built-in parallel ports. (thanks Brad Oldham)
git-svn-id: https://www.daphne-emu.com:9443/daphnesvn/branches/v_1_0@3012 c98d566b-fa28-4d4c-8a7b-932307ef5c1a

+++ src/globals.h
@@ -44,7 +44,7 @@ Uint8 frame_modifier = 0;	// what type of frame modifications to be used
 Uint8 joystick_detected = 0;	// 0 = no joystick, 1 = joystick present
 
 Uint8 realscoreboard = 0;	// 1 = external scoreboard is attached
-Uint8 rsb_port = 0;		// 0 = LPT1, 1 = LPT2 for real scoreboard
+unsigned int rsb_port = 0;		// 0 = LPT1, 1 = LPT2 or address of real scoreboard 
 
 unsigned int idleexit; // added by JFA for -idleexit
 unsigned char startsilent = 0;
